Implemented workqueue limiter settings on starting batch
