+++ Wabbajack/View Models/Compilers/VortexCompilerVM.cs
@@ -204,6 +204,9 @@ namespace Wabbajack
                     ModListWebsite = ModlistSettings.Website,
                     ModListReadme = ModlistSettings.ReadmeIsWebsite ? ModlistSettings.ReadmeWebsite : ModlistSettings.ReadmeFilePath.TargetPath,
                     ReadmeIsWebsite = ModlistSettings.ReadmeIsWebsite,
+                    ManualCoreLimit = Parent.MWVM.Settings.Performance.Manual,
+                    MaxCores = Parent.MWVM.Settings.Performance.MaxCores,
+                    TargetUsagePercent = Parent.MWVM.Settings.Performance.TargetUsage,
                 };
                 await ActiveCompilation.Begin();
             }
