+++ Wabbajack/View Models/Installers/MO2InstallerVM.cs
@@ -151,7 +151,12 @@ namespace Wabbajack
                 modList: Parent.ModList.SourceModList,
                 outputFolder: Location.TargetPath,
                 downloadFolder: DownloadLocation.TargetPath,
-                parameters: SystemParametersConstructor.Create());
+                parameters: SystemParametersConstructor.Create())
+            {
+                ManualCoreLimit = Parent.MWVM.Settings.Performance.Manual,
+                MaxCores = Parent.MWVM.Settings.Performance.MaxCores,
+                TargetUsagePercent = Parent.MWVM.Settings.Performance.TargetUsage,
+            };
 
             await Task.Run(async () =>
             {
